+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/playback/PlaybackState.kt
@@ -16,7 +16,6 @@ data class PlaybackState(
     val episodeUuid: String = "",
     val podcast: Podcast? = null,
     val fileMetadata: EpisodeFileMetadata? = null,
-    val embeddedArtworkPath: String? = null,
     val showNotesImageUrl: String? = null,
     val chapters: Chapters = Chapters(),
     val lastChangeFrom: String? = null,