+++ modules/services/servers/src/main/java/au/com/shiftyjelly/pocketcasts/servers/sync/NamedSettings.kt
@@ -56,7 +56,7 @@ data class ChangedNamedSettings(
     @field:Json(name = "autoSubscribeToPlayed") val autoSubscribeToPlayed: NamedChangedSettingBool? = null,
     @field:Json(name = "autoShowPlayed") val autoShowPlayed: NamedChangedSettingBool? = null,
     @field:Json(name = "autoPlayLastListUuid") val autoPlayLastSource: NamedChangedSettingString? = null,
-    @field:Json(name = "useEmbeddedArtworkGlobal") val useEmbeddedArtwork: NamedChangedSettingBool? = null,
+    @field:Json(name = "useEmbeddedArtworkGlobal") val useEpisodeArtwork: NamedChangedSettingBool? = null,
     @field:Json(name = "notificationActions") val notificationSettingActions: NamedChangedSettingString? = null,
     @field:Json(name = "playerShelfGlobal") val playerShelfItems: NamedChangedSettingString? = null,
     @field:Json(name = "showArtworkOnLockScreen") val showArtworkOnLockScreen: NamedChangedSettingBool? = null,
@@ -75,7 +75,6 @@ data class ChangedNamedSettings(
     @field:Json(name = "filesAfterPlayingDeleteCloudGlobal") val deleteCloudFilesAfterPlayback: NamedChangedSettingBool? = null,
     @field:Json(name = "filesAfterPlayingDeleteLocalGlobal") val deleteLocalFilesAfterPlayback: NamedChangedSettingBool? = null,
     @field:Json(name = "backgroundRefresh") val isPodcastBackgroundRefreshEnabled: NamedChangedSettingBool? = null,
-    @field:Json(name = "useRssArtwork") val useRssArtwork: NamedChangedSettingBool? = null,
 )
 
 @JsonClass(generateAdapter = true)