+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+7.62
+-----
+
+
 7.61
 -----
 *   New Features
@@ -6,8 +10,13 @@
 *   Updates:
     *   Display episode's RSS artwork in more places in the app.
         ([#1943](https://github.com/Automattic/pocket-casts-android/pull/1943))
+    *   Consolidate episode's embedded file artwork and RSS artwork into a single setting.
+        ([#1987](https://github.com/Automattic/pocket-casts-android/pull/1987))
     *   Add alphabetical sort order for podcast episodes.
-        ([#1968](https://github.com/Automattic/pocket-casts-android/pull/1968))
+        ([#1969](https://github.com/Automattic/pocket-casts-android/pull/1969))
+* Bug Fixes:
+    *   Subscription cancellation redirects now to a correct page.
+        ([#1973](https://github.com/Automattic/pocket-casts-android/pull/1973))
 
 7.60
 -----